requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
marshmallow==3.14.1
marshmallow-dataclass==8.5.3
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
tornado==6.1
typing-extensions==4.2.0
typing-inspect==0.7.1
tzdata==2022.1
tzlocal==4.2
